fix(sqcode): prevent integer overflow in large image scanning
The sq_scan_shape function used u32 accumulator variables (x_sum, y_sum, total_weight) when calculating weighted centers. For large images like the 739x739 pixel WiFi QR code, these accumulators overflowed when multiplying weight * x and summing across many pixels.

Changed accumulator types from u32 to u64 to support images up to 65535x65535 pixels with full 8-bit weights.

Added test case using pixel-wifi-sharing-qr-code.png to verify the fix.

Fixes crash introduced in 49f53300

+++ src/sqcode.rs
@@ -252,9 +252,9 @@ unsafe fn sq_scan_shape(img: *const zbar_image_t, dot: &mut Dot, start_x: i32, s
     // Calculate weighted center for dot
     let img_ref = &*img;
     let data = img_ref.data.as_ptr();
-    let mut x_sum = 0u32;
-    let mut y_sum = 0u32;
-    let mut total_weight = 0u32;
+    let mut x_sum = 0u64;
+    let mut y_sum = 0u64;
+    let mut total_weight = 0u64;
 
     for y in y0..(y0 + height) {
         for x in x0..(x0 + width) {
@@ -262,9 +262,9 @@ unsafe fn sq_scan_shape(img: *const zbar_image_t, dot: &mut Dot, start_x: i32, s
                 continue;
             }
             let idx = y as usize * img_ref.width as usize + x as usize;
-            let weight = (0xff - *data.add(idx)) as u32;
-            x_sum += weight * x;
-            y_sum += weight * y;
+            let weight = (0xff - *data.add(idx)) as u64;
+            x_sum += weight * x as u64;
+            y_sum += weight * y as u64;
             total_weight += weight;
         }
     }
